+++ SistemaTCC2/src/main/java/Models/Administrador.java
@@ -10,7 +10,7 @@ public class Administrador extends Usuario{
 	public Administrador() {
 	}
 	
-	public Administrador(Integer id) {
+	public Administrador(Long id) {
 		super.setId(id);
 	}
